+++ src/components/conversation/comment/commentTop.tsx
@@ -38,15 +38,9 @@ import { filterSwearWords } from "../../../lib/filterSwear";
 import UserModal from "./userModal";
 import { colors } from "../../../lib/css";
 import BlockedBtn from "./blockedBtn";
+import { useBlocked, useFold } from "../comment";
 
-export default function CommentTop(props: {
-    comment: Comment;
-    noStory?: boolean;
-    fold?: boolean;
-    setFold?: React.Dispatch<React.SetStateAction<boolean>>;
-    blocked?: boolean;
-    setBlocked?: React.Dispatch<React.SetStateAction<boolean | undefined>>;
-}) {
+export default function CommentTop(props: { comment: Comment; noStory?: boolean }) {
     const [open, setOpen] = useState(false);
     const [timeMode, setTimeMode] = useState<"short" | "long">("short");
     const [, setShareLink] = useShareLink();
@@ -62,9 +56,12 @@ export default function CommentTop(props: {
     const [user] = useUser();
     const [, setEditor] = useEditor();
     const [blockList] = useBlockList();
+    const [fold, setFold] = useFold();
+    const [blocked, setBlocked] = useBlocked();
+
     const cRoot = useCRoot();
 
-    const { comment, noStory, fold, setFold, blocked, setBlocked } = props;
+    const { comment, noStory } = props;
 
     const isOp = thread && thread.op.id === comment.user.id;
 
@@ -150,70 +147,70 @@ export default function CommentTop(props: {
     const moreList: (
         | { icon: JSX.Element; title: string; action: () => void }
         | undefined
-    )[] = useMemo(
-        () => [
-            (() => {
-                const clientIsOp = thread && user?.id === thread.op.id;
-                const pinned = thread?.pin?.id === comment.id;
-                if (clientIsOp || (user?.role === "admin" && pinned)) {
-                    const onError = (err: AxiosError<any>) => {
+    )[] = [
+        (() => {
+            const clientIsOp = thread && user?.id === thread.op.id;
+            const pinned = thread?.pin?.id === comment.id;
+            if (clientIsOp || (user?.role === "admin" && pinned)) {
+                const onError = (err: AxiosError<any>) => {
+                    setNotification({
+                        open: true,
+                        severity: "error",
+                        text: parseError(err),
+                    });
+                };
+                return {
+                    icon: <PushPinIcon />,
+                    title: `${pinned ? "Unpin" : "Pin"} Comment`,
+                    action: () => {
                         setNotification({
                             open: true,
-                            text: parseError(err),
+                            severity: "info",
+                            text: `${pinned ? "Unpinn" : "Pinn"}ing Comment...`,
                         });
-                    };
-                    return {
-                        icon: <PushPinIcon />,
-                        title: `${pinned ? "Unpin" : "Pin"} Comment`,
-                        action: () => {
-                            setNotification({
-                                open: true,
-                                text: `${pinned ? "Unpin" : "Pin"}ing Comment...`,
-                            });
-                            (pinned
-                                ? api.commentUnpin(threadId, comment.id)
-                                : api.commentPin(threadId, comment.id)
-                            )
-                                .then(() => {
-                                    setNotification({
-                                        open: true,
-                                        text: `Comment ${pinned ? "un" : ""}pinned!`,
-                                    });
-                                    setThread((thread) => {
-                                        if (!pinned && thread) thread.pin = comment;
-                                        else if (thread) delete thread.pin;
-                                        return thread;
-                                    });
-                                })
-                                .catch(onError);
-                        },
-                    };
-                }
-                return undefined;
-            })(),
-            {
-                icon: <FeedIcon className="!text-[19px]" />,
-                title: "Create thread",
-                action: () => {
-                    navigate(`/create?quote=${threadId}.${comment.id}`);
-                },
+                        (pinned
+                            ? api.threadUnpin(threadId)
+                            : api.threadPin(threadId, { cid: comment.id })
+                        )
+                            .then(() => {
+                                setNotification({
+                                    open: true,
+                                    severity: "success",
+                                    text: `Comment ${pinned ? "un" : ""}pinned!`,
+                                });
+                                setThread((thread) => {
+                                    if (!pinned && thread) thread.pin = comment;
+                                    else if (thread) delete thread.pin;
+                                    return thread;
+                                });
+                            })
+                            .catch(onError);
+                    },
+                };
+            }
+            return undefined;
+        })(),
+        {
+            icon: <FeedIcon className="!text-[19px]" />,
+            title: "Create thread",
+            action: () => {
+                navigate(`/create?quote=${threadId}.${comment.id}`);
             },
-            {
-                icon: <EditIcon className="!text-[19px]" />,
-                title: "Edit comment",
-                action: () => {
-                    if (user) setEditor({ open: true, edit: comment.comment });
-                    else
-                        navigate(
-                            `/users/login?continue=true&returnto=${encodeURIComponent(
-                                `${wholePath()}?c=${comment.id}`
-                            )}`
-                        );
-                },
+        },
+        {
+            icon: <EditIcon className="!text-[19px]" />,
+            title: "Edit comment",
+            action: () => {
+                if (user) setEditor({ open: true, edit: comment.comment });
+                else
+                    navigate(
+                        `/users/login?continue=true&returnto=${encodeURIComponent(
+                            `${wholePath()}?c=${comment.id}`
+                        )}`
+                    );
             },
-        ],
-        [comment, navigate, setEditor, setNotification, setThread, thread, threadId, user]
-    );
+        },
+    ];
 
     return (
         <Box