+++ src/components/conversation/functions/firstfetch.tsx
@@ -36,6 +36,7 @@ export default function useFirstFetch() {
         !notification.open &&
             setNotification({
                 open: true,
+                severity: "error",
                 text: parseError(err),
             });
         err?.response?.status === 404 && navigate("/404", { replace: true });
@@ -46,8 +47,8 @@ export default function useFirstFetch() {
             .then((data) => {
                 data.slink && setThread(data);
                 const historyIndex = history.findIndex((i) => i.id === threadId);
-                if (historyIndex && history[historyIndex].c < data.c) {
-                    history[historyIndex].c = data.c;
+                if (historyIndex && history[historyIndex].c < data.count) {
+                    history[historyIndex].c = data.count;
                     setHistory(history);
                     localStorage.setItem("history", JSON.stringify(history));
                 }