+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v4.4.0
+
+- add emotions list (opened by clicked the more button)
+- show users who expressed the emotion after clicking it in emotions list
+
+## v4.3.0
+
+- add emotions feature
+
 ## v4.2.0
 
 - star threads